% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_unhcr_pal.R
\name{display_unhcr_all}
\alias{display_unhcr_all}
\title{UNHCR color palettes viewer}
\usage{
display_unhcr_all(n = NULL, type = "all")
}
\arguments{
\item{n}{Number of different colors in the palette, minimum depending on the palette (2), maximum depending on the palette (5, 7 or 10)}

\item{type}{Type of the palette, can be "sequential", "diverging", "qualitative" or "all"}
}
\value{
No return value, called for side effects
}
\description{
Displays a set of the color palettes from UNHCR color
}
\examples{
library(unhcrthemes)
display_unhcr_all()
display_unhcr_all(type = "diverging")
display_unhcr_all(type = c("diverging", "qualitative", "sequential"))
display_unhcr_all(n = 3, type = "sequential")
display_unhcr_all(n = 5, type = "sequential")

}
