% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_unhcr_pal.R
\name{display_unhcr_pal}
\alias{display_unhcr_pal}
\title{UNHCR color palette viewer}
\usage{
display_unhcr_pal(n = NULL, name)
}
\arguments{
\item{n}{Number of different colors in the palette, minimum depending on palette (2), maximum depending on palette (5, 7 or 10)}

\item{name}{A palette name}
}
\value{
No return value, called for side effects
}
\description{
Displays a color palette from UNHCR brand color
}
\examples{
display_unhcr_pal(3, "pal_unhcr")

}
