% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idp.R
\docType{data}
\name{idp}
\alias{idp}
\title{Number of Internally Displaced People between 2010 and 2020.}
\format{
A data frame with 329 rows and 4 variables:
\describe{
  \item{year}{year}
  \item{country}{country}
  \item{sex}{sex of the idp}
  \item{idp}{total number of idp}
}
}
\source{
\url{http://www.unhcr.org/}
}
\usage{
idp
}
\description{
#' Number of Internally Displaced People in the World between
2010 and 2019.
}
\keyword{datasets}
