% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lato.R
\name{lato_installed}
\alias{lato_installed}
\title{Check if Lato font is installed in the system}
\usage{
lato_installed()
}
\value{
TRUE if Lato is installed in the system
}
\description{
Verify if you have Lato font installed in your system
}
