% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-unhcr-paged-generic.R
\name{unhcr_page_generic}
\alias{unhcr_page_generic}
\title{UNHCR Generic Paged Report}
\usage{
unhcr_page_generic(...)
}
\arguments{
\item{...}{Arguments passed to
\code{pagedown::\link[pagedown]{html_paged}}.}
}
\value{
An R Markdown output format.
}
\description{
Create a generic paged HTML report with UNHCR Branding that can be printed as PDF
}
