% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-unhcr-paged.R
\name{unhcr_paged}
\alias{unhcr_paged}
\title{UNHCR Paged Report}
\usage{
unhcr_paged(...)
}
\arguments{
\item{...}{Arguments passed to
\code{pagedown::\link[pagedown]{html_paged}}.}
}
\value{
An R Markdown output format.
}
\description{
Create a paged HTML report with UNHCR Branding that can be printed as PDF
}
