% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_plot.R
\name{unhcr_plot}
\alias{unhcr_plot}
\title{unhcr_plot}
\usage{
unhcr_plot(..., file = NULL)
}
\arguments{
\item{...}{urbn plot objects or grobs}

\item{file}{character, path of the file to save}

\item{heights}{relative heights of each object in the final plot}
}
\value{
one plot made from many grobs
}
\description{
unhcr_plot
}
