% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcr_save.R
\name{unhcr_save}
\alias{unhcr_save}
\title{Save ggplot2 using sensible defaults for better output}
\usage{
unhcr_save(plot = ggplot2::last_plot(), filename, dpi = 300, ...)
}
\arguments{
\item{plot}{ggplot2 object, plot to save. Defaults to last plot.}

\item{filename}{File name to create on disk.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster (png, jpeg and tiff) output types.}

\item{...}{extra param to passe to `ggsave`.}
}
\description{
Save ggplot2 using sensible defaults for better output
}
\examples{

\dontrun{
library(tidyverse)
library(unhcrthemes)

file <- tempfile("unhcrthemes_save", fileext = ".png")
cars |>
  ggplot(aes(speed, dist)) +
  geom_point() +
  theme_unhcr() |>
  unhcr_save(file)
}

}
