% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcrcolors.R
\docType{data}
\name{unhcrcolors}
\alias{unhcrcolors}
\title{UNHCR colors dataset}
\format{
A data frame with 11 rows and 13 variables:
* name: palette name
* type: palette type
* n2 to n11: list of the color values depending on the number of colors used (between two and twelve)
}
\usage{
unhcrcolors
}
\description{
A dataset containing the information about the CARTOcolor palettes.
Source: https://github.com/CartoDB/CartoColor/blob/master/cartocolor.js.
}
\keyword{datasets}
