% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unhcrthemes-package.R
\docType{package}
\name{unhcrthemes-package}
\alias{unhcrthemes}
\alias{unhcrthemes-package}
\title{unhcrthemes: UNHCR 'ggplot2' Theme and Colour Palettes}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

A 'ggplot2' theme and color palettes following the United Nations High Commissioner for Refugees (UNHCR) Data Visualization Guidelines recommendations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/unhcr-dataviz/unhcrthemes}
  \item \url{https://unhcr-dataviz.github.io/unhcrthemes/}
  \item Report bugs at \url{https://github.com/unhcr-dataviz/unhcrthemes/issues}
}

}
\author{
\strong{Maintainer}: Cédric Vidonne \email{cedric@vidonne.me}

Authors:
\itemize{
  \item Cédric Vidonne \email{cedric@vidonne.me}
  \item Ahmadou Dicko \email{mail@ahmadoudicko.com}
}

Other contributors:
\itemize{
  \item UNHCR [copyright holder]
}

}
\keyword{internal}
