% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-unhcr.R
\name{update_geom_font_defaults}
\alias{update_geom_font_defaults}
\title{Update matching font defaults for text geoms}
\usage{
update_geom_font_defaults(
  family = "Lato",
  face = "plain",
  size = 3.5,
  color = "#191919"
)
}
\arguments{
\item{family}{Font family, default "Lato"}

\item{face}{Font face, default "plain"}

\item{size}{Size}

\item{color}{Font color, default "#191919"}
}
\description{
Updates [ggplot2::geom_label] and [ggplot2::geom_text] font defaults
}
