% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_unhcr.R
\name{theme_unhcr}
\alias{theme_unhcr}
\title{UNHCR \link{ggplot2} theme}
\usage{
theme_unhcr(
  font_family = "Lato",
  font_size = 12,
  plot_title_size = rel(1.35),
  plot_title_margin = font_size,
  subtitle_size = font_size,
  subtitle_margin = font_size * 1.5,
  strip_text_size = rel(0.9),
  strip_text_face = "bold",
  strip_text_just = 0,
  caption_size = rel(0.7),
  caption_margin = font_size,
  axis_text = TRUE,
  axis_text_size = rel(0.9),
  axis_title = TRUE,
  axis_title_size = rel(0.9),
  axis = FALSE,
  axis_ticks = FALSE,
  grid = "XY",
  legend = TRUE,
  legend_text_size = rel(0.9),
  legend_title = FALSE,
  legend_title_size = rel(0.9),
  plot_margin = margin(font_size, font_size, font_size, font_size),
  plot_background = "white",
  void = FALSE
)
}
\arguments{
\item{font_family}{Base font family "Lato".}

\item{font_size}{Base font size in pt.}

\item{plot_title_size}{Plot title font size.}

\item{plot_title_margin}{Plot title bottom margin.}

\item{subtitle_size}{Plot subtitle font size.}

\item{subtitle_margin}{Plot subtitle bottom margin.}

\item{strip_text_size}{Facet label font size.}

\item{strip_text_face}{Facet label font face.}

\item{strip_text_just}{Facet label justification.}

\item{caption_size}{Caption font size.}

\item{caption_margin}{Caption top margin.}

\item{axis_text}{Logical \code{TRUE/FALSE} or \code{xy}. Turn on and off axis label.
Use \code{x} or \code{y} to have only the correspondent active.}

\item{axis_text_size}{Axis label font size.}

\item{axis_title}{Logical \code{TRUE/FALSE} or \code{xy}. Turn on and off axis title.
Use \code{x} or \code{y} to have only the correspondent active.}

\item{axis_title_size}{Axis title font size.}

\item{axis}{Logical \code{TRUE/FALSE} or \code{xy}. Turn on and off axis.
Use \code{x} or \code{y} to have only the correspondent active. Don't forget to review
the \code{expand} argument of the \verb{scale_x(or y)_...} to avoid spacing between
the axis and the baseline of the plot.}

\item{axis_ticks}{Logical \code{TRUE/FALSE} or \code{xy}. Turn on and off axis ticks.
Use \code{x} or \code{y} to have only the correspondent active.}

\item{grid}{Logical \code{TRUE/FALSE} or a combination of \code{XxYy}.
Turn on and off panel grids. Use \code{X} or \code{Y} for major
and \code{x} or \code{y} for minor.}

\item{legend}{Logical \code{TRUE/FALSE}. Turn on and off the legend.}

\item{legend_text_size}{Legend key font size.}

\item{legend_title}{Logical \code{TRUE/FALSE}. Turn on and off the legend title.}

\item{legend_title_size}{Legend title font size.}

\item{plot_margin}{Plot margin (specify with \code{ggplot2::margin()}).}

\item{plot_background}{Plot background, default "white".
Put "transparent" for no background color.}

\item{void}{Logical \code{TRUE/FALSE}. If \code{TRUE}, all grid lines, ticks
and axes are removed. Useful}
}
\value{
The theme style
}
\description{
\code{theme_unhcr} provides a basic \bold{UNHCR} theme
to use in \bold{ggplot2} commands.
}
\examples{
\dontrun{
library(ggplot2)

ggplot(mpg, aes(displ, hwy)) +
  geom_point() +
  theme_unhcr()
}
}
