% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-unhcr.R
\name{theme_unhcr}
\alias{theme_unhcr}
\title{UNHCR [ggplot2] theme following brand recommendations}
\usage{
theme_unhcr(
  base_family = "Lato",
  base_size = 10.5,
  legend_position = "top",
  plot_margin = ggplot2::margin(30, 30, 30, 30),
  grid = TRUE,
  axis_title = TRUE,
  axis = FALSE,
  ticks = FALSE
)
}
\arguments{
\item{base_family}{Base font family, default "Lato"}

\item{base_size}{Base font size}

\item{legend_position}{Position of the legend, default "top"}

\item{plot_margin}{Margin around the whole plot}

\item{grid}{Turn on and off the grids. "X" or "Y" for major and "x" or "y" for minor}

\item{axis_title}{Turn on and off axis titles. Use "X" or "Y" to have only the correspondent active}

\item{axis}{Turn on and off axis. Use "X" or "Y" to have only the correspondent active}

\item{ticks}{Turn on and off tick marks}
}
\description{
\code{theme_unhcr} provides a basic \bold{UNHCR} theme
to use in \bold{ggplot2} commands.
}
\examples{
\dontrun{
library(ggplot2)
library(showtext)

font_add("Lato", regular = "Lato-Regular.ttf", bold = "Lato-Bold.ttf", italic = "Lato-Italic.ttf")
showtext_auto()

# Simple scatterplot example

ggplot(datasets::iris, aes(x = Petal.Length, y = Petal.Width)) +
  geom_point() +
  labs(
    x = "Petal length", y = "Petal width",
    title = "Iris data ggplot2 scatterplot example",
    subtitle = "Just a simple plot to show the basic style of theme_unhcr",
    caption = "Data from datasets::iris"
  ) +
  theme_unhcr()
}

}
