% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_unhcr.R, R/theme_unhcr_map.R
\name{theme_unhcr}
\alias{theme_unhcr}
\alias{theme_unhcr_map}
\title{UNHCR [ggplot2] theme following brand recommendations}
\usage{
theme_unhcr(
  font_size = 9,
  font_family = "Lato",
  line_size = 0.5,
  rel_small = 8/9,
  rel_tiny = 7/9,
  rel_large = 12/9,
  grid = TRUE,
  axis = "x",
  ticks = FALSE,
  legend_title = FALSE
)

theme_unhcr_map(
  font_size = 14,
  font_family = "Lato",
  line_size = 0.5,
  rel_small = 12/14,
  rel_tiny = 11/14,
  rel_large = 16/14,
  legend_title = FALSE
)
}
\arguments{
\item{font_size}{Base font size}

\item{font_family}{Base font family, default "Lato"}

\item{line_size}{Base line size for axis. Gridlines are line_size/2}

\item{rel_small}{Relative size of small text (e.g., axis labels)}

\item{rel_tiny}{Relative size of tiny text (e.g., caption)}

\item{rel_large}{Relative size of large text (e.g., title)}

\item{grid}{Turn on and off the grids. "X" or "Y" for major and "x" or "y" for minor}

\item{axis}{Turn on and off axis. Use "X" or "Y" to have only the correspondent active}

\item{ticks}{Turn on and off ticks. Use "X" or "Y" to have only the correspondent active}

\item{legend_title}{Turn on and off title of the legend, default FALSE.}
}
\value{
The theme style

The theme style
}
\description{
\code{theme_unhcr} provides a basic \bold{UNHCR} theme
to use in \bold{ggplot2} commands.

\code{theme_unhcr} provides a basic \bold{UNHCR} theme
to use with map made using \bold{ggplot2}.
}
\examples{
\dontrun{
library(ggplot2)

ggplot(mpg, aes(displ, hwy)) +
geom_point() +
theme_unhcr()
}
}
