% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{unhcr_pal}
\alias{unhcr_pal}
\title{UNHCR color palettes}
\usage{
unhcr_pal(n = NULL, name, ...)
}
\arguments{
\item{n}{Number of different colors in the palette, minimum depending on palette (2), maximum depending on palette (5, 7 or 10)}

\item{name}{A palette name}

\item{...}{Additional arguments for [grDevices::colorRampPalette]}
}
\value{
A character vector
}
\description{
Creates a color palette from UNHCR branding
}
\examples{
unhcr_pal(3, "pal_main_qual")
unhcr_pal(3, "pal_blue_seq")
unhcr_pal(5, "pal_yellow_seq")
unhcr_pal(12, "pal_red_seq")
}
