% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_unhcr.R
\name{update_geom_font_defaults}
\alias{update_geom_font_defaults}
\title{Update matching font defaults for text geoms}
\usage{
update_geom_font_defaults(
  family = "Lato",
  face = "plain",
  size = 3.8,
  color = "#1a1a1a"
)
}
\arguments{
\item{family}{Font family, default "Lato".}

\item{face}{Font face, default "plain".}

\item{size}{font size in mm.}

\item{color}{Font color, default "#1a1a1a".}
}
\description{
Updates \link[ggplot2:geom_text]{ggplot2::geom_label} and \link[ggplot2:geom_text]{ggplot2::geom_text} font defaults
}
